Move conditional tulisp functions to a separate submodule
