+++ src/builtin/functions/mod.rs
@@ -1,9 +1,11 @@
 use crate::TulispContext;
 
+mod conditionals;
 mod functions;
 mod sequences;
 
 pub(crate) fn add(ctx: &mut TulispContext) {
+    conditionals::add(ctx);
     functions::add(ctx);
     sequences::add(ctx);
 }
